Manufacturing routes cycle time - save 1 decimal place (#251)
* most of back end for manufacturing routes

* resource builders that had been missed

* missing responses module register type definition for operations

* started integration tests and added empty constructor needed by entity framework

* removed add and update from module, added in missing work from resource builders and added include into repository for operations, set up links between routes and operations correctly in service db context, integration test for getting all fixed

* resource builder and integration test for getting by id fixed

* style cop fix

* stylecop fixes and uncommenting tests I wrongly commented out (mixed up folders)

* stylecop

* implemented updates etc for operations and updated modules to have routes for these for operations and resources

* unit tests for operation update and create added. Integration tests for operations added

* changed to using manufacturingId to identify operations since this is unique

* start of front end

* fixed merge mistake

* seems to now load correct stuff into redux store, not done anything with it yet though

* implemented basic view by id page, implemented search page client and backend + tests

* shared component implemented, table cells editable and back end updated to allow saving operations with the route

* tried and failed to get cancel to reset table data

* webpack changes

* method name changes

* most tests implemented properly for route and routes components, table fixed and basic tests implemented for it. Also removed duplicates from index reducer I had added by accident during merge

* tests updated and component tidied up

* tests cleaned up

* tests added and front end component changed to use shared component

* test fix

* added to app js menu, removed unnecessary () and formatted file

* permission checks implemented for update route

* working version with permission checks on update and add and tests fixed

* builder links tests added

* pr fixes

* remove unnecessary variable

* new service functionality to delete operations as part of updating routes, updates to components for item errors and to work with updates to inline editing table

* item error fixed in container

* added check that all operations contents had been filled out before save + pr fixes

* add missed item types

* fix to input is valid check for cycle time being 0

* make description and notes optional as per scott goodwin's request

* allow 0% as option

* redid checks for operations and gave error message specifying which columns don't meet criteria

* change cycle time to decimal and round to 1 decimal place

* updated the doubles to decimals

+++ src/Resources/ManufacturingOperationResource.cs
@@ -18,7 +18,7 @@
 
         public int SetAndCleanTime { get; set; }
 
-        public int CycleTime { get; set; }
+        public decimal CycleTime { get; set; }
 
         public int LabourPercentage { get; set; }
